double -> float.

+++ uboone/CRT/CRTProducts/CRTHit.hh
@@ -21,8 +21,8 @@ namespace crt {
 
     struct CRTHit{
       std::vector<uint8_t> feb_id;
-      std::map< uint8_t, std::vector<std::pair<int,double> > > pesmap;
-      double peshit;
+      std::map< uint8_t, std::vector<std::pair<int,float> > > pesmap;
+      float peshit;
      
       uint32_t ts0_s;
       uint16_t ts0_s_err;
@@ -41,12 +41,12 @@ namespace crt {
       uint16_t ts1_ns_err_corr;
 
       int plane;
-      double x_pos;
-      double x_err;
-      double y_pos;
-      double y_err;
-      double z_pos;
-      double z_err;
+      float x_pos;
+      float x_err;
+      float y_pos;
+      float y_err;
+      float z_pos;
+      float z_err;
       
       uint16_t event_flag;
       std::map< uint8_t, uint16_t > lostcpu_map;
